feat: 🎸 int => int64
int => int64

BREAKING CHANGE: 🧨 int => int64

✅ Closes: int => int64

+++ stores/gormx/scopes/paginate/model.go
@@ -2,21 +2,21 @@ package paginate
 
 // Pagination 分页
 type Pagination struct {
-	Page     int   `json:"page"`
-	PageSize int   `json:"page_size"`
+	Page     int64 `json:"page"`
+	PageSize int64 `json:"page_size"`
 	Total    int64 `json:"total"`
 	Rows     any   `json:"rows"`
 	Extend   any   `json:"extend,omitempty"`
 }
 
-func (p *Pagination) Offset() int {
+func (p *Pagination) Offset() int64 {
 	if p.Page <= 0 {
 		p.Page = 1
 	}
 	return (p.Page - 1) * p.PageSize
 }
 
-func (p *Pagination) Limit() int {
+func (p *Pagination) Limit() int64 {
 	if p.PageSize < 10 {
 		p.PageSize = 10
 	}
@@ -26,10 +26,10 @@ func (p *Pagination) Limit() int {
 	return p.PageSize
 }
 
-func (p *Pagination) GetPage() int {
+func (p *Pagination) GetPage() int64 {
 	return p.Page
 }
 
-func (p *Pagination) GetPageSize() int {
+func (p *Pagination) GetPageSize() int64 {
 	return p.PageSize
 }
